+++ packages/context/api.d.ts
@@ -1,23 +1,4 @@
-export {
-  getContext,
-  isContext,
-  getCDNBaseUrl,
-  initializeManifestData,
-  initializeUI5YamlData,
-  findManifestPath,
-  findYamlPath,
-  getManifestDetails,
-  getUI5Manifest,
-  getUI5Yaml,
-  getYamlDetails,
-  reactOnCdsFileChange,
-  reactOnUI5YamlChange,
-  reactOnManifestChange,
-  reactOnXmlFileChange,
-  reactOnPackageJson,
-  cache,
-  DEFAULT_I18N_NAMESPACE,
-} from "./src/api";
+export * from "./src/api";
 
 export type {
   Context,