+++ packages/fe/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 4.0.31
+
+### Patch Changes
+
+- Updated dependencies [a82d24c]
+  - @ui5-language-assistant/context@4.0.28
+  - @ui5-language-assistant/xml-views-completion@4.0.28
+
 ## 4.0.30
 
 ### Patch Changes