+++ packages/vscode-ui5-language-assistant-bas-ext/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 4.0.68
+
+### Patch Changes
+
+- a82d24c: fix: reference
+
 ## 4.0.67
 
 ### Patch Changes